+++ client/src/modules/global-event-handlers.js
@@ -24,4 +24,6 @@ const globalEventHandlers = () => {
   });
 };
 
-export { globalEventHandlers };
+module.exports = {
+  globalEventHandlers,
+};