+++ client/src/modules/world.js
@@ -1,7 +1,6 @@
-import { light } from './light.js';
-import { skybox } from './skybox.js';
-import { keyboardControls } from './keyboard-controls.js';
-import { themeMusic } from './theme-music.js';
+const { light, skybox, keyboardControls } = require('ebabel');
+
+const { themeMusic } = require('./theme-music');
 
 // Default renderer clear color.
 const _color = 0x0e0727;
@@ -12,6 +11,7 @@ const world = (input) => {
   const {
     THREE,
     THREEx,
+    dataStore,
     color = _color,
     opacity = _opacity
   } = input;
@@ -31,10 +31,15 @@ const world = (input) => {
   camera.rotation.set(...dataStore.player.state.rotation);
 
   // Setup keyboard controls.
-  const controls = keyboardControls();
+  const controls = keyboardControls(dataStore);
 
   // Setup main theme music.
-  const music = themeMusic({ camera });
+  const music = themeMusic({
+    THREE,
+    camera,
+    volume: dataStore.defaultVolume,
+    url: 'assets/music/ambient2-nautilus.mp3',
+  });
 
   // Setup main renderer for WebGL graphics.
   const renderer = new THREE.WebGLRenderer({ antialias: true });
@@ -49,7 +54,12 @@ const world = (input) => {
   light({ THREE, scene });
 
   // Skybox.
-  skybox(scene);
+  skybox({
+    THREE,
+    scene,
+    directions:  ['ft', 'bk', 'up', 'dn', 'rt', 'lf']
+      .map((direction) => `../assets/whirlpool/large-files/whirlpool_${direction}.jpg`)
+  });
 
   return {
     clock,
@@ -62,4 +72,6 @@ const world = (input) => {
   };
 };
 
-export { world };
+module.exports = {
+  world,
+};